postgres_to_es/requirements.txt
python-dotenv==0.15.0
psycopg2==2.8.6
# Elasticsearch 7.x
elasticsearch>=7.0.0,<8.0.0
loguru>=0.5.3,<0.6.0
